Improvements to the cursor width. #1632
If the user is on Windows, the system-wide cursor width is respected. It can be changed in the accessibility settings part of the control panel ("ease of access -> optimise visual display").
If on Linux/OS X, by now the cursor width will be of 1px again, and an option to change it can be added in the future.