Disable the option to disable QGtkStyle on Qt 5. Fixes #1610
We had to disable QGtkStyle on Qt 4 by default because of several bugs (#777, #988, #964), but this is no longer a problem on Qt 5.
In fact, disabling QGtkStyle on Qt 5 leaves KVIrc in a non-functional state, where menu entries are no longer visible and everything looks like you used a time machine to go back to 1995.