client: Revamp user mode support
serverinfo: Add all user modes with descriptions to each IRCd's applicable
class. Create tables for user mode requirements to determine whether or not
the user mode can be applied to the user.
coreactions: Display information on each user mode in the drop down menu
for user mode selection. Check if the user mode can be set with the users
current permissions, if it cannot then gray it out.