Don't cache nsis and enchant.7z
One is preinstalled, another is not used anymore.

As a side effect, fix #1865 by mere fact of touching .appveyor.yml
because something got corrupted in our build cache.
Probably the corrupted thing was nsis, because it was claiming to pack
locale/ directory, but resulting installer didn't have it, as shown by
7zip.
I suspected that version of preinstalled nsis was different from cached
nsis, but they both were 2.46