Implement `znc.in/self-message` IRCv3 capability
This capability is used to indicate that the client supports receiving
messages previously sent by the user. In ZNC, it is required for proper
playback of private message conversations.

More information: http://wiki.znc.in/Query_buffers#Self_messages

The implementation is fairly straight-forward. In the relevant places,
we switch the method used to identify if a message is a private message
or belongs in a channel - instead of comparing the target to our
nickname, we check if it starts with one of the server-announced
CHANTYPES characters, which has the additional benefit of avoiding a
race condition when changing nicknames.

Possible forward improvements:

- Currently, if the first message received for a window is a
  self-message, we do not call setTarget. As far as I've seen, the
  effect is purely cosmetic (the target is not added to the member list,
  which is hidden by default anyway).

- CTCP code for each CTCP message type needs to be manually updated to
  handle self-messages. This patch only addresses ACTION CTCPs.