Shut down cleanly on SIGTERM and SIGINT
SIGTERM is the standard way on POSIX for the OS to tell an application
to close. It can be sent when the computer is shutting down or the user
is logging off, as well as by the user, directly, or indirectly (e.g. by
using a task manager).

SIGINT is sent when the user presses Ctrl+C in a terminal window in
which the program is running.

In both of these cases, KVIrc should shut down cleanly, as if the user
closed the main window. This implies flushing logs.

By implementing SIGTERM and SIGINT handling on POSIX, this patch fixes
the log files not being flushed on receiving such a signal, thus causing
the last few lines to be lost.