NickServWidget: Use raw NickServ instead of msg NickServ
IRCds have aliases in their configuration to route messages to services
specifically. In other words, instead of doing /msg nickserv the IRCd
treats it like /msg NickServ!NickServ@services.generic.net.
This is to aid in the security that sensitive message will get routed to
the service, in the even that if services are down and a malicious user takes
over the nick, messages using /nickserv will not be sent to that user.