fix: hide comments in issues in new posts
this happens because comments can only have one ( open and ) close bracket. so ``` [//]:# (this is a comment (and it) shouldnt show up)``` shows up unless the inner brackets are escaped like [//]:# (this is a comment \(and it\) shouldnt show up) or theres no inner brackets or ```<!-- this -->``` is used.

or something else, no idea

this should fix the issue.