Avatars: Another "rethink defaults" change
The default behavior now for Avatars are that they are disabled by
default. Instead, when the user enables avatars, they will not show
up directly within the user list. The KVIrc logo will instead be
displayed next to the user within the user list, indicating that the
user has extra KVIrc metadata available. This replaces gender icons
for now.