Remove gzipped logs as an option
Despite the fact gzip'd log data is written with max compression,
it is done by appending independently-compressed data to the
relevant archive on every flush or log stop - this basically kills
the compression as there is little duplication in the small amount
of data logged. Over time, this produces archives that are much
larger than the data stored (therefore making the 'compression'
pointless).

gzip is still used outside of logging in KVIrc (and to read 'old'
logs), so is not removed.

Fixes 'gzip compression of logs on flushing/log stopping in many
cases results in larger archives than log data' / #2353