Update Desktop Entry file This commit adds well-formed values for all keys defined in version 1.5 of the [Freedesktop.org Desktop Entry Specification][1], published on 2020-04-27, adding: - the ability for KVIrc to be designated by all XDG-compliant Linux   Desktop Environments (GNOME, KDE et al.) as a (default) handler for   irc/irc6/ircs protocol URIs within all installed applications, - support for the Startup Notification Protocol, - a context menu entry in DE graphical application launchers for opening   a new session/instance/window of KVIrc when one or more are already in   use, and - registration of KVIrc as the default handler for the 'text/kvc' and   'text/kvs' Media (MIME) Type definitions already installed by the   several circulating vendor packages for Linux
This effectively upstreams the [patch implemented by Kubuntu][2] on
2009-03-08 and later merged by Debian on 2011-05-07 into the official
Debian kvirc packaging repository to add keywords to the launcher entry
which help users to find the entry when searching for all IRC/Chat apps
installed on their system.

The desktop entry file itself is renamed to
`cmake/net.kvirc.KVIrc.desktop.cmake` in the source repository to comply
with above-referenced DES v1.5, which requires that such files use a
valid "D-Bus well-known name" as their basename (which is also the
"desktop file ID" used many other XDG-compliant interfaces), derived, as
suggested, by appending the mixed-case application name to the reverse
DNS entry of the official homepage. The CMake build configuration is
altered to further append the major version number to the mixed-case
application name prior to installation, as advised by the specification.

Signed-off-by: Peter J. Mello <admin@petermello.net>

1: https://specifications.freedesktop.org/desktop-entry-spec/latest/
2: https://salsa.debian.org/qt-kde-team/extras/kvirc/-/blob/93f8469f20a7a36a9956f3dc0e3d1c87178486a8/debian/patches/kubuntu_01_fix_desktop_entry.patch